refactor(leads): extract HTTP layer to src/web/leads.rs + small fixes
- Move all lead-capture HTTP handlers, forms, helpers and the background
  purge/notifier loop from src/web/mod.rs (~580 lines) to src/web/leads.rs.
  Phone validation helpers stay in mod.rs since they cross-cut the booking
  flow, not lead capture. `pub use leads::run_lead_purge_loop` preserves
  the path main.rs imports.
- Promote impersonation_ctx, sidebar_context, CsrfForm to pub(crate) so
  leads.rs can call them.
- Drop the now-dead POST /dashboard/event-types/{slug}/lead-capture route
  + toggle_event_type_lead_capture handler + LeadCaptureToggleForm struct.
  The toggle is part of the main event-type form since phase 2.
- Fix phone validation to count chars instead of bytes so a 64-char phone
  with non-ASCII formatting marks isn't rejected by the server while
  passing the HTML `maxlength=64` attribute on the client.
- Replace the chars().take(10) hack on lead_capture_acknowledged_at with
  a proper NaiveDateTime parse + format, falling back to the raw value if
  the column ever stores a non-SQLite-default format.
- Add Fluent keys for every new English string introduced in phase 2
  (event-type-form-phone-*, event-type-form-lead-capture-*,
  admin-legal-*, admin-lead-*). EN only; other locales fall back to EN at
  runtime per the existing i18n contract.
- Document the BookForm/4-handler duplication as a standalone plan in
  notes/booking-handler-duplication.md, with the test invariants a
  unified BookingFlow refactor would unlock.
