+++ templates/event_type_form.html
@@ -452,13 +452,13 @@
     </div>
 
     <div class="form-group" style="margin-top: 0.75rem;">
-      <label for="collect_phone">Phone number</label>
+      <label for="collect_phone">{{ t("event-type-form-phone-label") }}</label>
       <select id="collect_phone" name="collect_phone" style="padding: 0.625rem 0.75rem; border: 1px solid var(--border); border-radius: var(--radius); font-size: 0.925rem; font-family: inherit; background: var(--surface); color: var(--text); width: 100%;">
-        <option value="0" {% if form_collect_phone == 0 %}selected{% endif %}>Don't ask</option>
-        <option value="1" {% if form_collect_phone == 1 %}selected{% endif %}>Ask, but optional</option>
-        <option value="2" {% if form_collect_phone == 2 %}selected{% endif %}>Ask, and required</option>
+        <option value="0" {% if form_collect_phone == 0 %}selected{% endif %}>{{ t("event-type-form-phone-off") }}</option>
+        <option value="1" {% if form_collect_phone == 1 %}selected{% endif %}>{{ t("event-type-form-phone-optional") }}</option>
+        <option value="2" {% if form_collect_phone == 2 %}selected{% endif %}>{{ t("event-type-form-phone-required") }}</option>
       </select>
-      <span class="hint" style="font-size: 0.8rem; color: var(--text-muted); display: block; margin-top: 0.25rem;">When collected, the phone number is included in the calendar event description and on host notification emails.</span>
+      <span class="hint" style="font-size: 0.8rem; color: var(--text-muted); display: block; margin-top: 0.25rem;">{{ t("event-type-form-phone-help") }}</span>
     </div>
 
     <div class="form-group" style="margin-top: 0.75rem;">
@@ -605,23 +605,23 @@
      ON pops a confirm() and stamps the acknowledgement as a hidden field. The
      server keeps the ack gate as a fallback when JS is disabled. -->
 <div class="card" style="margin-top: 1.25rem;">
-  <h2 style="margin-bottom: 0.5rem;">Lead capture</h2>
+  <h2 style="margin-bottom: 0.5rem;">{{ t("event-type-form-lead-capture-heading") }}</h2>
   <p class="subtitle" style="margin-bottom: 1rem;">
-    Record what guests type into this event type's booking form before they submit, so you can follow up if they abandon the booking.
+    {{ t("event-type-form-lead-capture-intro") }}
     {% if not lead_capture_global_enabled %}
-      <br><strong>Disabled by your administrator.</strong> Ask them to enable it from the admin panel before this toggle takes effect.
+      <br><strong>{{ t("event-type-form-lead-capture-admin-disabled") }}</strong>
     {% endif %}
   </p>
   <label style="display: flex; align-items: flex-start; gap: 0.625rem; cursor: pointer; padding: 0.75rem; border: 1px solid var(--border); border-radius: var(--radius);">
     <input type="checkbox" name="lead_capture" id="lead-capture-enabled" {% if form_lead_capture %}checked{% endif %} style="accent-color: var(--accent); width: 1.125rem; height: 1.125rem; margin-top: 0.125rem;">
     <span>
-      <strong style="font-size: 0.925rem;">Enable lead capture for this event type</strong>
-      <div style="color: var(--text-muted); font-size: 0.825rem;">Guests' name, email, phone and notes are auto-saved as they type. Captured rows are auto-purged after {{ lead_capture_retention_days }} days. Enabling will ask you to confirm that bookers are informed.</div>
+      <strong style="font-size: 0.925rem;">{{ t("event-type-form-lead-capture-toggle") }}</strong>
+      <div style="color: var(--text-muted); font-size: 0.825rem;">{{ t("event-type-form-lead-capture-help", days=lead_capture_retention_days) }}</div>
     </span>
   </label>
   <input type="hidden" name="lead_capture_acknowledged" id="lead-capture-acknowledged" value="{% if lead_capture_acknowledged_at %}on{% endif %}">
   {% if lead_capture_acknowledged_at %}
-    <div style="margin-top: 0.5rem; font-size: 0.8rem; color: var(--text-muted);">Acknowledged on {{ lead_capture_acknowledged_at }}.</div>
+    <div style="margin-top: 0.5rem; font-size: 0.8rem; color: var(--text-muted);">{{ t("event-type-form-lead-capture-acknowledged", date=lead_capture_acknowledged_at) }}</div>
   {% endif %}
 </div>
 {% endif %}
@@ -645,7 +645,7 @@
     var initiallyOn = toggle.checked;
     form.addEventListener('submit', function(e) {
       if (toggle.checked && ackField.value !== 'on') {
-        var msg = "By enabling lead capture, you confirm that bookers are informed their input is recorded. Continue?";
+        var msg = {{ t("event-type-form-lead-capture-confirm") | tojson }};
         if (window.confirm(msg)) {
           ackField.value = 'on';
         } else {
